TLM ROS C/C++ interface based on a bug fix of original FATODE
In original FATODE's ROS_tlm_f90_Integrator, routine lapack_free
does not free saved module member "ip"(for pivoting in
lapack solver), this causes allocation error in lapack_init
when the integrator is run the second time in a same
simulation. A fix is made to LS_Solver.F90 to determine if
ip should be allocated based on its status.